% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{bootMarkov}
\alias{bootMarkov}
\title{bootMarkov}
\usage{
bootMarkov(
  obs_adj_mat,
  obs_node_freq,
  directed = TRUE,
  loops = TRUE,
  ci = 0.95,
  R = 1000
)
}
\arguments{
\item{<x>}{text file name}
}
\description{
This function returns a numeric sequence corresponding to a text
file, suitable for RQA use.
}
\examples{
bootMarkov(ts)

}
\keyword{clean}
\keyword{text}
