% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guastello.R
\docType{data}
\name{guastello}
\alias{guastello}
\title{Sample categorical time series}
\format{
A character vector with 20 elements.
}
\source{
Guastello, S. (2011). Orbital decomposition: Identification of
dynamical patterns in categorical data. in S. Guastello & R. Gregson (Eds.)
Nonlinear dynamical systems analysis for the behavioral sciences using
real data. Boca Raton, FL: CRC Press.
}
\usage{
guastello
}
\description{
A dataset containing a ficticious sequence of categorical data.
}
\keyword{datasets}
