% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{makeMarkov}
\alias{makeMarkov}
\title{makeMarkov}
\usage{
makeMarkov(ts)
}
\arguments{
\item{<x>}{text file name}
}
\description{
This function returns a numeric sequence corresponding to a text
file, suitable for RQA use.
}
\examples{
makeMarkov(ts)

}
\keyword{clean}
\keyword{text}
