% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpoweRlaw.R
\name{orbde}
\alias{orbde}
\title{orbde}
\usage{
orbde(data_seq)
}
\arguments{
\item{<data_seq>}{character or integer time series}
}
\description{
This function returns the result of orbital decomposition. All of the
metrics associated with orbital decomposition are returned in a list.
}
\examples{
orbde(guastello)

}
\keyword{decomposition}
\keyword{orbital}
