% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDSToolkit.R
\name{prepareText}
\alias{prepareText}
\title{prepareText}
\usage{
prepareText(x)
}
\arguments{
\item{<x>}{text file name}
}
\description{
This function cleans text data for use in RQA.
}
\examples{
prepareText(letitbe)

}
\keyword{clean}
\keyword{text}
