% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDSToolkit.R
\name{time_burstiness}
\alias{time_burstiness}
\title{time_burstiness}
\usage{
time_burstiness(times, min_iet = NULL)
}
\arguments{
\item{<ts>}{numeric vector}

\item{<min_iet>}{minimum interevent sequence spacing}
}
\description{
This function returns the burstiness coefficient for a series
of events. The series must be time-based; see event_burstiness() to find
the burstiness of codes in a event-series.
}
\examples{
event_burstiness(guastello, 1)
event_burstiness(guastello)
time_burstiness()

}
\references{
Kim, E.-K., & Jo, H.-H. (2016). Measuring burstiness for finite
event sequences. Physical Review E, 94(3), 032311.
https://doi.org/10.1103/PhysRevE.94.032311
}
\keyword{burstiness}
