% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDSToolkit.R
\name{bootMarkov}
\alias{bootMarkov}
\title{bootMarkov}
\usage{
bootMarkov(
  obs_adj_mat,
  obs_node_freq,
  directed = TRUE,
  loops = TRUE,
  ci = 0.95,
  R = 1000
)
}
\arguments{
\item{<obs_adj_mat>}{observed adjacency matrix}

\item{<obs_node_freq>}{observed node frequencies}

\item{<directed>}{is the network directed (TRUE) or not (FALSE)}

\item{<loops>}{are there self-loops in the matrix?}

\item{<ci>}{confidence interval level}

\item{<R>}{number of replications}
}
\description{
This function returns a matrix of the same dimension as
obs_adj_mat, with values of {-1,0,1}. Values of -1 (+1) indicate that the
corresponding entry in obs_adj_mat has an entry that is less (more) than
would be expected by the observed node frequencies and outside the
bootstrapped confidence interval matrix created from the observed node
frequencies.
}
\details{
This function returns a numeric sequence corresponding to a text
file, suitable for RQA use.
}
\examples{
bootMarkov()

}
\keyword{clean}
\keyword{text}
