% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDSToolkit.R
\name{event_burstiness}
\alias{event_burstiness}
\title{event_burstiness}
\usage{
event_burstiness(ts, min_iet = NULL)
}
\arguments{
\item{<ts>}{character (or logicial) vector}

\item{<min_iet>}{minimum interevent sequence spacing}
}
\description{
This function returns the burstiness coefficient for each of the
unique codes in a series. The series must be event-based; see
time_burstiness() to find the burstiness of codes in a time-series.
}
\examples{
event_burstiness(guastello, 1)
event_burstiness(guastello)

}
\references{
Kim, E.-K., & Jo, H.-H. (2016). Measuring burstiness for finite
event sequences. Physical Review E, 94(3), 032311.
https://doi.org/10.1103/PhysRevE.94.032311
}
\keyword{burstiness}
