% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpoweRlaw.R
\name{IPL_fit}
\alias{IPL_fit}
\title{IPL_fit}
\usage{
IPL_fit(ts, C = 1)
}
\arguments{
\item{<ts>}{character time series}
}
\description{
This function returns the burstiness coefficient for each of the
unique codes in the time series.
}
\examples{
IPL_fit(guastello, C)

}
\keyword{burstiness}
