% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buRst.R
\name{time_burstiness}
\alias{time_burstiness}
\title{time_burstiness}
\usage{
time_burstiness(times, min_iet = NULL)
}
\arguments{
\item{<ts>}{series of numeric times of events}

\item{<min_iet>}{minimum interevent time}
}
\description{
This function returns the burstiness coefficient for a series of events.
The series must be time-based; see event_burstiness() to find the
burstiness of codes in a event-series.
}
\examples{
time_burstiness()

}
\keyword{burstiness}
